+++ src/openapi/streaming/streaming-websocket.spec.js
@@ -14,7 +14,6 @@ import WebSocketTransport from './connection/transport/websocket-transport';
 import Connection from './connection/connection';
 import '../../test/mocks/math-random';
 import Streaming from './streaming';
-import log from '../../log';
 import * as constants from './connection/constants';
 import mockAuthProvider from '../../test/mocks/authProvider';
 
@@ -27,7 +26,6 @@ describe('openapi Streaming', () => {
     let connectionSlowCallback;
     let startCallback;
     let receivedCallback;
-    let errorCallback;
     let authProvider;
     let mockConnection;
     let subscriptionUpdateSpy;
@@ -57,9 +55,6 @@ describe('openapi Streaming', () => {
         mockConnection.received.mockImplementation((callback) => {
             receivedCallback = callback;
         });
-        mockConnection.error.mockImplementation((callback) => {
-            errorCallback = callback;
-        });
         mockConnection.connectionSlow.mockImplementation((callback) => {
             connectionSlowCallback = callback;
         });
@@ -85,10 +80,6 @@ describe('openapi Streaming', () => {
                 connectionSlowCallback = callback;
             },
         );
-
-        Connection.prototype.setErrorCallback.mockImplementation((callback) => {
-            errorCallback = callback;
-        });
 
         global.$ = {
             connection: jest.fn().mockReturnValue(mockConnection),
@@ -521,7 +512,7 @@ describe('openapi Streaming', () => {
         });
     });
 
-    describe('websockt events', () => {
+    describe('websocket events', () => {
         let streaming;
         beforeEach(() => {
             streaming = new Streaming(
@@ -538,10 +529,41 @@ describe('openapi Streaming', () => {
             connectionSlowCallback();
             expect(connectionSlowSpy.mock.calls.length).toEqual(1);
         });
-        it('handles connection error events', () => {
-            jest.spyOn(log, 'error');
-            errorCallback('error details');
-            expect(log.error.mock.calls.length).toEqual(1);
+    });
+
+    describe('network issues', () => {
+        it('calls through', () => {
+            const streaming = new Streaming(
+                transport,
+                'testUrl',
+                authProvider,
+                {},
+            );
+
+            const subscription = streaming.createSubscription(
+                'root',
+                '/test/test',
+                {},
+                subscriptionUpdateSpy,
+                subscriptionErrorSpy,
+            );
+
+            jest.spyOn(streaming.connection, 'onSubscribeNetworkError');
+            jest.spyOn(streaming.connection, 'onOrphanFound');
+
+            expect(() => {
+                subscription.onNetworkError();
+            }).not.toThrow();
+
+            expect(
+                streaming.connection.onSubscribeNetworkError,
+            ).toBeCalledTimes(1);
+
+            expect(() => {
+                streaming.orphanFinder.onOrphanFound(subscription);
+            }).not.toThrow();
+
+            expect(streaming.connection.onOrphanFound).toBeCalledTimes(1);
         });
     });